+++ xsolla-store-sdk/src/main/java/com/xsolla/android/store/entity/request/cart/UpdateItemBody.kt
@@ -1,3 +1,3 @@
 package com.xsolla.android.store.entity.request.cart
 
-data class UpdateItemBody(val quantity: Int)
+data class UpdateItemBody(val quantity: Long)
